# Whole program dead code elimination