--- a/.gitattributes
+++ b/.gitattributes
@@ -24,6 +24,11 @@
 *.png binary
 *.tfm binary
 
+# 'union' merge driver just unions textual content in case of conflict
+#   http://krlmlr.github.io/using-gitattributes-to-avoid-merge-conflicts/
+/.mailmap                merge=union
+/Changes                 merge=union
+
 # No header for text files (would be too obtrusive).
 *.md                     ocaml-typo=missing-header
 README*                  ocaml-typo=missing-header