--- a/.travis-ci.sh
+++ b/.travis-ci.sh
@@ -58,7 +58,7 @@ EOF
           OCAML_NATIVE_TOOLS=true &&
         make all &&
         make install)
-    git clone git://github.com/ocaml/camlp4
+    git clone git://github.com/ocaml/camlp4 -b 4.05
     (cd camlp4 &&
      ./configure --bindir=$PREFIX/bin --libdir=$PREFIX/lib/ocaml \
        --pkgdir=$PREFIX/lib/ocaml && \