--- a/Changes
+++ b/Changes
@@ -192,6 +192,10 @@ Next version (4.05.0):
   a short description in overviews.
   (Florian Angeletti)
 
+- PR#7488: ocamldoc, wrong Latex output for variant types
+  with constructors without arguments.
+  (Florian Angeletti, report by Xavier Leroy)
+
 - GPR#848: ocamldoc, escape link targets in HTML output
   (Etienne Millon, review by Gabriel Scherer, Florian Angeletti and
   Daniel Bünzli)