--- a/Changes
+++ b/Changes
@@ -388,6 +388,11 @@ Next version (4.05.0):
 - GPR#1019: Fix fatal error in Flambda mode "[functions] does not map set of
   closures ID" (Pierre Chambart, code review by Mark Shinwell and Leo White)
 
+- GPR#1042: Fix escaping of command-line arguments in
+  Unix.create_process{,_env} under Windows.  Arguments with tabs should now
+  be received verbatim by the child process.
+  (Nicolas Ojeda Bar, Andreas Hauptmann review by Xavier Leroy)
+
 ### Toplevel and debugger:
 
 - PR#7060, GPR#1035: Print exceptions in installed custom printers