--- a/Changes
+++ b/Changes
@@ -26,6 +26,8 @@ Next version (4.05.0):
   passed to the new functions are handled by the garbage collector.
   (Gabriel Scherer, review by Mark Shinwell, request by Immanuel Litzroth)
 
+- GPR#953 (was MPR#385): Add caml_startup_exn (Mark Shinwell)
+
 - GPR#891: Use -fno-builtin-memcmp when building runtime with gcc.
   (Leo White)