I propose in this PR to add another link mode to the `ocamlopt` compiler with flambda enabled.

There is a new `-lto` option to the compiler to mark when a file should export sufficient information for the link. When all the `.cmx` and `.cmxa` files of a project are built with this option, it is possible to link with it too. When linking, all the flambda informations are concatenated and go through a dead code elimination pass that removes all unreferenced symbols (but keep effets) and build a new big object file containing the whole program.
Of course, this prevent using dynlink in this program as the modules referenced by the loaded module might have been eliminated. It may be possible to provide a mode where some interfaces are requested, and only those would be available for a dynlinked module but this is not implemented yet.
There are no optimization specific to whole program applied yet.

Outside of the undynlinkability, the other drawbacks are:
* cmxa files are really larger. (stdlib goes from 15kB to 1.9MB). Those files didn't contain much,
  for this mode, they must contain the whole code of the included modules since the compiler
  might not have access to the cmx files while linking.
* cmx files are not that much bigger (less than 10% on the stdlib), probably quite more in -Oclassic

Overall compilation time does not change significantly, but link time of course increases a lot:

without `-lto`
```
$ time ./ocamlopt.opt -dtimings -g -nostdlib -I stdlib -I otherlibs/dynlink  -ccopt "-Wl,-E" -o ocamlc.opt   compilerlibs/ocamlcommon.cmxa compilerlibs/ocamlbytecomp.cmxa   driver/main.cmx -cclib "-lm  -ldl -lcurses -lpthread" 

real	0m0.565s
user	0m0.464s
sys	0m0.088s
```

with `-lto`
```
$ time ./ocamlopt.opt -dtimings -g -nostdlib -I stdlib -I otherlibs/dynlink  -ccopt "-Wl,-E" -lto -o ocamlc.opt   compilerlibs/ocamlcommon.cmxa compilerlibs/ocamlbytecomp.cmxa   driver/main.cmx -cclib "-lm  -ldl -lcurses -lpthread"

real	0m8.243s
user	0m7.912s
sys	0m0.316s

all: 5.752s
flambda(concatenate)(link): 0.096s
flambda(remove_unused_program_constructs)(link): 0.512s
flambda(backend)(link): 0.452s
cmm(link): 0.604s
compile_phrases(link): 2.992s
selection(link): 0.228s
comballoc(link): 0.020s
cse(link): 0.180s
deadcode(link): 0.088s
spill(link): 0.312s
split(link): 0.136s
liveness(link): 0.264s
regalloc(link): 1.088s
linearize(link): 0.024s
scheduling(link): 0.004s
emit(link): 0.400s
assemble(link): 0.008s
```

The effect is not wonderful on the compiler itself: the size of ocamlc.opt decreases by ~10% (there is not much dead code there), but on some extreme examples we can get quite a lot. There are still some cases where this does not eliminate as much as expected.

This patch is based on https://github.com/ocaml/ocaml/pull/602 only the commits after https://github.com/ocaml/ocaml/commit/88c2c8ce796ccb9ff7beb54c7c03971ba43841a9 (Also remove linking hack for bytecode) are relevant. This other PR is needed to allow removing unneeded toplevel modules.

Note the `-lto` or 'link time optimization' is quite badly named. Please suggest a better option.