@let-def 2016-06-08
Rather than making cmxa largely redundant with cmx files, couldn't we get rid of cmxa files (by e.g, allowing to have link options in cmx files)?
I am not asking for this to be mandatory (that would break compatibility very badly), maybe just to make this workflow viable, now that cmx files are needed for flambda to operate properly.