@lpw25 2016-06-08
@let-def I agree and have a patch that does this as part of a larger "namespaces" proposal.