@xavierleroy 2016-06-08
Concerning @let-def's suggestion, it's exactly what Caml Light did back in the days.  Then, someone objected on the basis of the following scenario: you have a library "foo" containing two modules, "foo_aux" and "foo".  With the proposed approach, you first compile foo.ml, obtaining foo.cmx, then build the library foo.cmx, overwriting the previous foo.cmx file...

Also, .cmx files describe .o files while .cmxa files describe .a files.  What are you proposing?  .cmx files that describe .a files?  Get rid of .a files?