@let-def 2016-06-08
Yes, offer a workflow without .a files. Libraries add one more level of names, I would like to be able to do without if possible.
The linker could receive all objects (with existing tools, it's mostly a matter of putting all .o files in the META description).