@lpw25 2016-06-08
I had assumed @let-def was suggesting allowing to use a directory filled with `.cmx` files in place of a `.cmxa`. You already need such a directory for cross-module inlining so why not use it for linking.