@samoht 2016-06-08
@chambart I'm curious to see the gain that this could have on some MirageOS binaries. Is there a way to turn `-lto` by default on the whole switch (and if yes, could this be added as a new experimental opam compiler?)