@chambart 2016-06-08
@let-def why not, but this probably does not fit in this PR (that is already large enouth). Currently beside link information, there is also no way to tell using only cmx files that a file should be linked only if used. This matter if there is some initialization code. We could of course also add a flag to the cmx files to tell something like that.

@samoht you can use `export OCAMLPARAM=lto=1,_` or use something similar to the the hack in: https://github.com/ocaml/opam-repository/pull/6115 (which by the way was not propagated to the current default 4.03 compiler as I just noticed...)