@chambart 2016-06-08
The last patch adds a minimal optimization round when linking to allow to remove some more references to toplevel modules: when something like:

module A
```
initialize_symbol a (call ...)
initialize_symbol camlA (a.(0))
```

module B
```
let v = A.camlA.(0)
```
When concatenating A and B, B still maintain a reference to camlA but it could have been redirected to A.a. This usually does not matter since reaching the value in a field of a symbol or another has the same performance cost, hence there is no information propagated in the cmx file to know that A.camlA.(0) is an alias of A.a.(0).

It now matters. This alias is something that inline_and_simplify knows about, so running it after concatenation allows to remove all references to camlA.

In practice that can appear for instance if we use stdin that force the whole pervasive module to stay alive. A quite extreme example, with this pass and sufficiently aggressive inlining option (`-O3 -inline-max-unroll 4`), the complete code of the program:

```
Printf.printf "Hello world %i\n" 3
```

Is in clambda (this is after un-anf, hence constants are not shown)

```
un-anf (init_code):
(seq
  (caml_register_named_value "camlPervasives__Pccall_arg_1389"
    "camlPervasives__Pmakeblock_1388")
  (caml_ml_open_descriptor_in 0)
  (setfield_ptr(init) 0 "camlPervasives__Pccall_1585"
    (caml_ml_open_descriptor_out 1))
  (caml_ml_open_descriptor_out 2)
  (setfield_ptr(init) 0 "camlPervasives__exit_function_1549"
    (makemutable 0 "camlPervasives__flush_all_582_closure"))
  (caml_register_named_value "camlPervasives__Pccall_arg_1460"
    "camlPervasives__do_at_exit_1213_closure")
  (setfield_ptr(init) 0 "camlTest__Pccall_282"
    (caml_format_int "camlCamlinternalFormat__const_string_10802" 3))
  (caml_ml_output (field 0 "camlPervasives__Pccall_1585")
    "camlTest__Pmakeblock_arg_34" 0 12)
  (setfield_ptr(init) 0 "camlTest__Pccall_arg_280"
    (string.length (field 0 "camlTest__Pccall_282")))
  (caml_ml_output (field 0 "camlPervasives__Pccall_1585")
    (field 0 "camlTest__Pccall_282") 0 (field 0 "camlTest__Pccall_arg_280"))
  (caml_ml_output_char (field 0 "camlPervasives__Pccall_1585") 10)
  (setfield_ptr(init) 0 "camlStd_exit__simplify_fv_22"
    (field 0 (field 0 "camlPervasives__exit_function_1549")))
  (apply (field 0 "camlStd_exit__simplify_fv_22") 0a) 0a)

un-anf (camlPervasives__iter_586):
(if param_588/1262
  (let
    (sequence_591/1264
       (try (caml_ml_flush (field 0 param_588/1262)) with exn_592/1267 0a))
    (apply* camlPervasives__iter_586  (field 1 param_588/1262)))
  0a)


un-anf (camlPervasives__do_at_exit_1213):
(apply (field 0 (field 0 "camlPervasives__exit_function_1549")) 0a)

un-anf (camlPervasives__flush_all_582):
(apply* camlPervasives__iter_586  (caml_ml_out_channels_list 0a))
```

and the constants are
```
Pervasives.camlPervasives__Pmakeblock_arg_1387: "index out of bounds"
CamlinternalFormat.camlCamlinternalFormat__const_string_10802: "%i"

Pervasives.camlPervasives__Pccall_arg_1389: "Pervasives.array_bound_error"

Test.camlTest__Pmakeblock_arg_34: "Hello world "

Pervasives.camlPervasives__Pccall_arg_1460: "Pervasives.do_at_exit"

Pervasives.camlPervasives__Pmakeblock_1388: block(0,"caml_exn_Invalid_argument","camlPervasives__Pmakeblock_arg_1387")
```

Notice that the majority of the code is related to do_at_exit