@samoht 2016-06-08
@chambart any change that you could fix the 4.03 description and add your new compiler to opam-repository? :-) (it not, I'll try to do this next week)