@alainfrisch 2016-06-08
Did you consider an actual "whole program optimizer" based on flambda?  I can imagine such a compiler loading e.g. `.cmt` files (produced "quickly" by `ocamlc`, or by `ocamlopt` in non-flambda mode) and doing all the compilation+linking globally.  The closed world assumption opens many more opportunities for optimizations (including tweaks to the representation of values if they cannot be observed from C).  This would support quick compilation (including with ocamlopt) for reasonably fast edit/compile cycle but also local testing, while allowing full optimization mode for building critical executables.