@mshinwell 2016-06-09
(Also, I'm not sure "-lto" is badly named.  It describes what is going on and is nearly the same as the corresponding GCC option for the same thing.)